Move pgBoss setup to a separate file pgBoss.js
